requirements.txt
fastapi
uvicorn
pytubefix
pydantic
python-multipart
aiosqlite 
loguru==0.7.2
python-dotenv